+++ src/Plugin/views/field/BulkForm/DependentContentBulkForm.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @file
+ * Contains \Drupal\dependent_content\Plugin\views\field\BulkForm\DependentContentBulkForm
+ */
+
+namespace Drupal\dependent_content\Plugin\views\field\BulkForm;
+
+
+use Drupal\system\Plugin\views\field\BulkForm;
+
+/**
+ * Defines a dependent content operations bulk form element.
+ *
+ * @ViewsField("dependent_content_bulk_form")
+ */
+class DependentContentBulkForm extends BulkForm {}